glossary/requirements.txt
google-cloud-translate
google-cloud-storage